refactor: split inventory monolith, group commands/, honest file-size limits
Architecture hygiene pass:
- Split scan/inventory.ts (1082 lines, one giant function-bag) into a focused
  scan/inventory/ module: constants, paths, walk, manifests, hints, git, and an
  index.ts orchestrator. Each concern is now independently readable/testable;
  largest piece is 243 lines.
- Group the 8 root-level command + review-packet files under src/commands/, so
  the layering reads commands/ (CLI) -> scan/ (I/O) -> ghost-core/ (model).
  review-packet (domain-ish, the odd one out at root) now lives with its review
  command. Fixed skill-command's bundle-path URL for the new depth.
- Delete all three file-size EXCEPTIONS: two were stale lies (cli.ts granted 580
  for 161 actual; fingerprint-commands 1135 for 178) and inventory no longer
  needs one. Every file now passes the honest 500-line default.

All green: 109 tests, full check, zero file-size exceptions.
