+++ packages/ghost/src/commands/command-discovery.ts
